+++ stable/xml.cpp
@@ -80,15 +80,10 @@ qRT
 qRE
 }
 
-static status__ SkipSpaces_(
-  _flow___ &Flow,
-  str::dString *Spaces = NULL)
+static status__ SkipSpaces_(_flow___ &Flow)
 {
 	while ( !Flow.EndOfFlow() && isspace( Flow.View() ) )
-    if ( Spaces != NULL )
-      Spaces->Append(Flow.Get());
-    else
-      Flow.Get();
+    Flow.Get();
 
 	if ( Flow.EndOfFlow() )
 		return sUnexpectedEOF;
@@ -504,7 +499,6 @@ static status__ GetAttribute_(
 	_flow___ &Flow,
 	entities_handling__ EntitiesHandling,
 	str::string_ &Name,
-	str::dString &Between,  // Chars between the attribute name and the attribute value, including the '=' sign.
 	bso::sChar &Delimiter,
 	str::string_ &Value)
 {
@@ -515,14 +509,12 @@ static status__ GetAttribute_(
 		return sUnexpectedCharacter;
 	}
 
-	HANDLE( SkipSpaces_(Flow, &Between) );
+	HANDLE( SkipSpaces_(Flow) );
 
 	if ( Flow.Get() != '=' )
 		return sMissingEqualSign;
 
-	Between.Append('=');
-
-  HANDLE( SkipSpaces_(Flow, &Between) );
+  HANDLE( SkipSpaces_(Flow) );
 
 	Delimiter = Flow.Get();
 
@@ -815,10 +807,9 @@ qRB
 			case t_Undefined:
 				AttributeName_.Init();
 				AttributeDelimiter_ = delimiter::sUndefined;
-				AttributeBetween_.Init();
 				_Value.Init();
 
-				HANDLE( GetAttribute_( _Flow, _EntitiesHandling, AttributeName_, AttributeBetween_, AttributeDelimiter_, _Value ) );
+				HANDLE( GetAttribute_( _Flow, _EntitiesHandling, AttributeName_, AttributeDelimiter_, _Value ) );
 
 				if ( _Tags.IsEmpty() )
 					qRFwk();
@@ -1080,7 +1071,7 @@ status__ xml::Parse(
 	status__ Status = s_Undefined;
 qRH
 	parser___ Parser;
-	str::string TagName, AttributeName, AttributeBetween, Value;
+	str::string TagName, AttributeName, Value;
 	bso::sChar AttributeDelimiter = delimiter::sUndefined;
 	bso::bool__ Stop = false;
 	xml::dump Dump;
@@ -1090,12 +1081,11 @@ qRB
 	while ( !Stop ) {
 		TagName.Init();
 		AttributeName.Init();
-		AttributeBetween.Init();
 		AttributeDelimiter = delimiter::sUndefined;
 		Value.Init();
 		Dump.PurgeData();
 
-		switch ( Parser.Parse( TagName, AttributeName, AttributeBetween, AttributeDelimiter, Value, Dump, Status ) ) {
+		switch ( Parser.Parse( TagName, AttributeName, AttributeDelimiter, Value, Dump, Status ) ) {
 		case tProcessingInstruction:
 			Stop = !Callback.XMLProcessingInstruction( Dump );
 			break;
@@ -1106,10 +1096,10 @@ qRB
 			Stop = !Callback.XMLStartTagClosed( TagName, Dump );
 			break;
 		case tAttribute:
-			Stop = !Callback.XMLAttribute( TagName, AttributeName, AttributeBetween, AttributeDelimiter, Value, Dump );
+			Stop = !Callback.XMLAttribute( TagName, AttributeName, AttributeDelimiter, Value, Dump );
 			break;
 		case tSpecialAttribute:
-			Stop = !Callback.XMLSpecialAttribute( TagName, AttributeName, AttributeBetween, AttributeDelimiter, Value, Dump );
+			Stop = !Callback.XMLSpecialAttribute( TagName, AttributeName, AttributeDelimiter, Value, Dump );
 			break;
 		case tValue:
 		case tCData: